requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.3
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
